refactor : 예외처리 변경에 따른 수정 및 패키지 구조 변경에 따른 수정
