+++ src/main/java/model/state/BowlingState.java
@@ -1,6 +1,6 @@
 package model.state;
 
-import model.frame.Score;
+import model.frame.score.Score;
 import model.pin.PinCount;
 import model.state.status.Status;
