+++ src/main/java/model/state/running/RunningState.java
@@ -1,6 +1,7 @@
 package model.state.running;
 
-import model.frame.Score;
+import model.frame.score.NotCountScore;
+import model.frame.score.Score;
 import model.state.BowlingState;
 import model.state.status.Status;
 
@@ -17,6 +18,6 @@ public abstract class RunningState implements BowlingState {
 
     @Override
     public Score getScore() {
-        throw new IllegalStateException("투구 실행 중에는 점수를 확인 할 수 없습니다.");
+        throw new NotCountScore("투구 실행 중에는 점수를 확인 할 수 없습니다.");
     }
 }
